Rename Ansi_Processor to Processing_Ansi, created new Abstract Base class so we can reuse the base class for Message editor data parsing also.
