+++ src/form_manager.cpp
@@ -2,7 +2,7 @@
 
 #include "model-sys/menu.hpp"
 #include "forms/form_system_config.hpp"
-#include "ansi_processor.hpp"
+#include "processor_ansi.hpp"
 #include "session_data.hpp"
 #include "logging.hpp"
 
@@ -83,13 +83,13 @@ void FormManager::startupForm(form_ptr form)
     m_ansi_bot = m_common_io.readinAnsi(m_form.back()->m_ansi_bot);
 
     // Calc Top Rows, get Ending Y Position
-    ansi->parseAnsiScreen((char *)m_ansi_top.c_str());
+    ansi->parseTextToBuffer((char *)m_ansi_top.c_str());
     int top_rows = ansi->getYPosition();
 
     ansi->clearScreen();
 
     // Calc Bottom Rows
-    ansi->parseAnsiScreen((char *)m_ansi_bot.c_str());
+    ansi->parseTextToBuffer((char *)m_ansi_bot.c_str());
     int bot_rows = ansi->getMaxRowsUsedOnScreen();
 
     ansi->clearScreen();
@@ -354,7 +354,7 @@ std::string FormManager::processMidFormTemplate(const std::string &screen)
 
     // Clear Codemap.
     std::vector<MapType>().swap(code_map);
-    ansi_process->parseAnsiScreen((char *)output_screen.c_str());
+    ansi_process->parseTextToBuffer((char *)output_screen.c_str());
 
     // Return with no clear screen, since this is a mid ansi.
     return ansi_process->getScreenFromBuffer(false);
