+++ src/processor_ansi.cpp
@@ -1,4 +1,4 @@
-#include "ansi_processor.hpp"
+#include "processor_ansi.hpp"
 #include "model-sys/structures.hpp"
 #include "common_io.hpp"
 #include "logging.hpp"
@@ -21,37 +21,6 @@
 #include <fstream>
 #include <vector>
 #include <sstream>
-
-
-AnsiProcessor::AnsiProcessor(int term_height, int term_width)
-    : m_ansi_output("")
-    , m_is_screen_cleared(false)
-    , m_is_line_wrapping(false)
-    , m_position(0)
-    , m_y_position(1)
-    , m_number_lines(term_height)
-    , m_characters_per_line(term_width)
-    , m_x_position(1)
-    , m_max_x_position(1)
-    , m_max_y_position(1)
-    , m_center_ansi_output(false)
-    , m_saved_cursor_x(1)
-    , m_saved_cursor_y(1)
-    , m_attribute(0)
-    , m_saved_attribute(0)
-    , m_saved_foreground(FG_WHITE)
-    , m_saved_background(BG_BLACK)
-    , m_foreground_color(FG_WHITE)
-    , m_background_color(BG_BLACK)
-    , m_screen_pixel()
-{
-    m_screen_buffer.reserve(m_number_lines * m_characters_per_line);
-    m_screen_buffer.resize(m_number_lines * m_characters_per_line);
-}
-
-AnsiProcessor::~AnsiProcessor()
-{
-}
 
 
 /**
@@ -475,36 +444,36 @@ std::string AnsiProcessor::screenBufferParse()
         {
             case 1:
                 // Then process and display the lightbars.
+            {
+                int pull_id = 0;
+                std::stringstream ss;
+                ss.str(my_matches.m_code.substr(1, 2));
+                ss >> pull_id;
+
+                if(ss.fail())
                 {
-                    int pull_id = 0;
-                    std::stringstream ss;
-                    ss.str(my_matches.m_code.substr(1, 2));
-                    ss >> pull_id;
-
-                    if(ss.fail())
-                    {
-                        ss.clear();
-                        ss.ignore();
-                        break;
-                    }
-
-                    // Grab the highlight color from the second sequence %##.
-                    m_screen_buffer[my_matches.m_offset].selected_attribute =
-                        m_screen_buffer[my_matches.m_offset+3].attribute;
-
-                    m_screen_buffer[my_matches.m_offset].selected_foreground =
-                        m_screen_buffer[my_matches.m_offset+3].foreground;
-
-                    m_screen_buffer[my_matches.m_offset].selected_background =
-                        m_screen_buffer[my_matches.m_offset+3].background;
-
-                    // tear out the y and x positions from the offset.
-                    m_pull_down_options[pull_id] = m_screen_buffer[my_matches.m_offset];
-
                     ss.clear();
                     ss.ignore();
+                    break;
                 }
-                break;
+
+                // Grab the highlight color from the second sequence %##.
+                m_screen_buffer[my_matches.m_offset].selected_attribute =
+                    m_screen_buffer[my_matches.m_offset+3].attribute;
+
+                m_screen_buffer[my_matches.m_offset].selected_foreground =
+                    m_screen_buffer[my_matches.m_offset+3].foreground;
+
+                m_screen_buffer[my_matches.m_offset].selected_background =
+                    m_screen_buffer[my_matches.m_offset+3].background;
+
+                // tear out the y and x positions from the offset.
+                m_pull_down_options[pull_id] = m_screen_buffer[my_matches.m_offset];
+
+                ss.clear();
+                ss.ignore();
+            }
+            break;
 
             default:
                 break;
@@ -675,7 +644,7 @@ void AnsiProcessor::clearScreen()
  * @brief Parses screen data into the Screen Buffer.
  * @return
  */
-void AnsiProcessor::parseAnsiScreen(char *buff)
+void AnsiProcessor::parseTextToBuffer(char *buff)
 {
     if(strlen(buff) == 0)
         return;
