+++ xrm-unittest/xrm-unittest.project
@@ -29,6 +29,10 @@
   <Description/>
   <Dependencies/>
   <VirtualDirectory Name="src">
+    <File Name="../src/processor_base.hpp"/>
+    <File Name="../src/processor_ansi.hpp"/>
+    <File Name="../src/processor_ansi.cpp"/>
+    <File Name="processor_ansi_ut.cpp"/>
     <File Name="../src/logging.hpp"/>
     <File Name="../src/encoding.hpp"/>
     <File Name="../src/encoding.cpp"/>
@@ -67,9 +71,6 @@
     <File Name="session_io_ut.cpp"/>
     <File Name="../src/session_io.cpp"/>
     <File Name="../src/session_io.hpp"/>
-    <File Name="ansi_processor_ut.cpp"/>
-    <File Name="../src/ansi_processor.cpp"/>
-    <File Name="../src/ansi_processor.hpp"/>
     <File Name="users_dao_ut.cpp"/>
     <File Name="../src/session_manager.cpp"/>
     <File Name="../src/session_manager.hpp"/>
